requirements.txt
django
djangorestframework
psycopg2
ffmpeg-python
celery
